feat(mondo): router layout
