+++ .vscode/settings.json
@@ -14,4 +14,5 @@
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   },
+  "editor.tabCompletion": "on"
 }
