+++ apps/mondo/package.json
@@ -11,8 +11,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "codefee-kit": "0.0.42-alpha.23",
+    "lodash-es": "^4.17.21",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-is": "^18.2.0",
+    "react-router-dom": "^6.4.3",
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
     "@types/react": "^18.0.22",
@@ -23,6 +28,7 @@
     "modelite": "workspace:*",
     "tsconfig": "workspace:*",
     "typescript": "^4.6.4",
-    "vite": "^3.2.0"
+    "vite": "^3.2.0",
+    "vite-tsconfig-paths": "^3.5.2"
   }
 }
