+++ apps/mondo/src/layouts/RootLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RootLayout';
