+++ apps/mondo/src/layouts/SuspenseLoader/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SuspenseLoader';
+export * from './SuspenseLoader';
