+++ apps/mondo/src/pages/Settings/Settings.tsx
@@ -1,0 +1,7 @@
+import { CircularProgress } from 'codefee-kit';
+
+const Settings = () => {
+  return <CircularProgress />;
+};
+
+export default Settings;
