+++ apps/mondo/src/pages/Todo/index.ts
@@ -1,0 +1,1 @@
+export { default } from './todo';
