+++ apps/mondo/src/typings/index.ts
@@ -1,0 +1,1 @@
+export * from './react-ext';
