+++ apps/mondo/src/typings/react-ext.ts
@@ -1,0 +1,3 @@
+import { FC, PropsWithChildren } from 'react';
+
+export type FcWithChildren<T = unknown> = FC<PropsWithChildren<T>>;
